+++ default.nix
@@ -1,20 +1,19 @@
-{ pkgs ? (import <nixpkgs> {}).pkgs }:
+{ pkgs ? import <nixpkgs> { }
+}:
 
 let
-  # List any extra packages you want available while your package is
-  # building or while in a nix shell:
-  extraPackages = with pkgs; [ ];
-
   # Helpful if you want to override any Haskell packages:
-  haskell = pkgs.haskellPackages;
+  overrides = self: super: {
+    http-client = if super ? http-client_0_6_2
+      then super.http-client_0_6_2
+      else super.http-client;
+  };
+
+  # Apply the overrides from above:
+  haskell = pkgs.haskellPackages.override (orig: {
+    overrides = pkgs.lib.composeExtensions
+      (orig.overrides or (_: _: {})) overrides; });
 in
 
 # Load the local nix file and use the overrides from above:
-haskell.callPackage ./themoviedb.nix {
-  mkDerivation = { buildTools ? []
-                 , ...
-                 }@args:
-    haskell.mkDerivation (args // {
-      buildTools = buildTools ++ extraPackages;
-    });
-}
+haskell.callPackage ./themoviedb.nix { }